Fix remaining rand/zune-jpeg/sha2 API breakage from the patch+minor bump
- rand_distr 0.6 requires Zipf's parameter to be a Float, not u64 -- build
  the distribution over f64 and cast when sampling
- zune-jpeg 0.5 requires its reader to implement Seek, which a bare &[u8]
  no longer satisfies -- wrap it in zune_core's ZCursor as the crate's own
  docs recommend
- sha2 0.11 switched Sha256::finalize()'s output to a type that no longer
  implements LowerHex -- hex-encode the digest bytes manually

+++ spikes/den/src/gen.rs
@@ -137,7 +137,7 @@ pub fn generate_catalog(opts: &GenOptions) -> Vec<Asset> {
     }
 
     // Zipf-distributed keyword vocabulary: a handful of hot keywords, a long tail.
-    let keyword_zipf = Zipf::new((LOCATIONS.len() + SUBJECTS.len()) as u64, 1.2).unwrap();
+    let keyword_zipf = Zipf::new((LOCATIONS.len() + SUBJECTS.len()) as f64, 1.2).unwrap();
     let all_keywords: Vec<&str> = LOCATIONS.iter().chain(SUBJECTS.iter()).copied().collect();
 
     let mut assets = Vec::with_capacity(opts.asset_count as usize);
